requirements.txt
flask
ultralytics
pillow
numpy
werkzeug
flask-cors